Implementing Dataview - Add DataviewService module and separate common logic
